decimal not compatible with json, using float

+++ lambda/get-game/coordinates.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 def from_numbers_to_strings(game):
     game['startPos']['long'] = str(game['startPos']['long'])
     game['startPos']['lat'] = str(game['startPos']['lat'])
@@ -12,10 +9,10 @@ def from_numbers_to_strings(game):
 
 
 def from_strings_to_numbers(game):
-    game['startPos']['long'] = Decimal(game['startPos']['long'])
-    game['startPos']['lat'] = Decimal(game['startPos']['lat'])
+    game['startPos']['long'] = float(game['startPos']['long'])
+    game['startPos']['lat'] = float(game['startPos']['lat'])
     for path in game['paths']:
         for point in path:
-            point['long'] = Decimal(point['long'])
-            point['lat'] = Decimal(point['lat'])
+            point['long'] = float(point['long'])
+            point['lat'] = float(point['lat'])
     return game
